[e131] Use uint16_t for consumer refcount
Struct pads to 4 bytes either way due to alignment, so this is free.

+++ esphome/components/e131/e131.h
@@ -24,7 +24,7 @@ struct E131Packet {
 
 struct UniverseConsumer {
   uint16_t universe;
-  uint8_t consumers;
+  uint16_t consumers;
 };
 
 class E131Component : public esphome::Component {
